Use composite primary key to enable #14
Following https://github.com/hassox/guardian_db/pull/14 for this to work you'd need a composite primary key and not just an extra unique index.

Currently `jti` is already set as primary key and (at least in PostgreSQL) that means it will be unique. Inserting an entry with a non unique `jti` to the table will just cause error so another unique index doesn't do anything and the whole lookup by extra field change seem to be superfluous, it should work as intended with the composite key.